Вынос модуля Timeit и настройка среды
