+++ Q4.py
@@ -1,5 +1,5 @@
 # coding=utf-8
-import time
+import timeit
 """
 Задача 4. Наибольшее произведение-палиндром
 
@@ -8,15 +8,6 @@ import time
 
 Найдите самый большой палиндром, полученный умножением двух трехзначных чисел.
 """
-def timeit(method):
-    def timed(*args, **kw):
-        ts = time.time()
-        result = method(*args, **kw)
-        te = time.time()
-        print('%r  %2.2f ms' % (method.__name__, (te - ts) * 1000))
-        return result
-    return timed
-
 def is_poly(num):
     start = str(num)
     end = len(start) - 1
@@ -26,7 +17,6 @@ def is_poly(num):
 
     return True
 
-@timeit
 def poly_detected():
     max = 0
     for i in range(1000, 100, -1):
@@ -37,7 +27,7 @@ def poly_detected():
     return max
 
 
-#print(poly_detected())
+print(poly_detected())
 # 205.15 ms
